requirements.txt
requests
lxml
click
fpdf
pillow
fpdf2